ml_services/report_service/requirements.txt
grpcio
grpcio-tools
matplotlib
python-dotenv

geopandas
contextily
shapely
xyzservices

google-generativeai
python-dotenv